早口言葉が正しく言えているか、信頼度が選択しているレベル以上かを判定できるようにした

+++ app/src/main/java/com/example/twist4english/LevelSlideFragment.kt
@@ -24,7 +24,7 @@ class LevelSlideFragment : Fragment() {
             linearLayout.findViewById<TextView>(R.id.level_title).text =
                 it.getString(TITLE)
 
-            val requiredConfidence = it.getInt(CONFIDENCE_SCORE).toString()
+            val requiredConfidence = it.getFloat(CONFIDENCE_SCORE).toString()
             val message =
                 "means you have to get $requiredConfidence points or more for each tongue twister!"
             linearLayout.findViewById<TextView>(R.id.level_description).text = message
